requirements.txt
greenlet==2.0.1
pyaes==1.6.1
pyasn1==0.4.8
python-dotenv==0.21.0
rsa==4.9
SQLAlchemy==1.4.42
Telethon==1.25.4
